refactor(buffer): Factor out movement into separate module and cleanup
